initial interactive button
